third_party/WebKit/LayoutTests/http/tests/websocket/url-attribute.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">

description(".url attribute should work, previously supported .URL should not");

var url = "ws://127.0.0.1:8880/echo";
var ws = new WebSocket(url);

shouldBeEqualToString("ws.url", url);
shouldBeUndefined("ws.URL");

</script>
</body>
</html>
