third_party/WebKit/LayoutTests/http/tests/websocket/url-attribute-expected.txt

CONSOLE WARNING: 'WebSocket.URL' is deprecated. Please use 'WebSocket.url' instead.
Both .URL and .url attributes should work (for compatibility reasons).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.url is "ws://127.0.0.1:8880/echo"
PASS ws.URL is "ws://127.0.0.1:8880/echo"
PASS ws.URL === ws.url is true
PASS successfullyParsed is true

TEST COMPLETE
